backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1

# LLM (direct SDK + LangChain — all routed through LiteLLM proxy)
openai==1.59.6
httpx==0.28.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.62

# Auth (Google OAuth + email/password)
authlib==1.3.2
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0

# Vector store
chromadb==1.5.8

# Google APIs (Sheets, Drive)
google-api-python-client==2.151.0
google-auth==2.49.2
gspread==6.2.1

# File handling
pypdf==5.1.0
pillow==11.1.0

# Dev / test / lint
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.4
